backend/requirements.txt
Django>=5.0.0,<6.0.0
djangorestframework>=3.14.0
django-tenants>=3.5.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
gunicorn>=21.2.0
whitenoise>=6.6.0
channels[daphne]>=4.1.0
channels-redis>=4.2.0
redis>=5.0.0